Cambios en dominio ,cambio de BD
Se adapto el dominio a la nueva BD , y se intento un boton para agregar un nuevo medico a la DB.

+++ Dominio/Sede.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
    public class Sede
     {
-        public int ídSede { get; set; }
+        public long ídSede { get; set; }
         public string NombreSede { get; set; }
 
     }
